Move llama-server cleanup code to a separate file
