+++ src-tauri/src/core/utils/extensions/inference_llamacpp_extension/mod.rs
@@ -1,1 +1,2 @@
 pub mod server;
+pub mod cleanup;
